# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Galileo API
  version: "0.3"
  description: |
    Observability + LLM gateway. Auth: session cookie (login) or `Authorization: Bearer glt_…` personal token.
    OTLP ingest is on :4317/:4318 with a project API key (`glk_…`); the gateway is under `/gw/`.
servers:
  - url: http://localhost:8080
security:
  - bearer: []
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer }
...
paths:
...
  /api/projects/{id}/gateway/routes: { get: { summary: Gateway routes, parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { "200": { description: OK } } }, post: { summary: Create a route (targets, budget, guardrails, semantic cache, smart routing), parameters: [{ name: id, in: path, required: true, schema: { type: string } }], responses: { "200": { description: OK } } } }
...